requirements.txt
Flask==3.1.0
google-cloud-aiplatform==1.90.0
google-generativeai==0.5.0
numpy==2.2.5
pandas==2.2.3
Pillow==11.2.1
